Fix overflow for very small probabilities in BinomialScanner.

+++ src/main/scala/ru/ifmo/onell/util/BinomialScanner.scala
@@ -3,17 +3,17 @@ package ru.ifmo.onell.util
 import java.util.Random
 
 abstract class BinomialScanner {
-  def offset(rng: Random): Int
+  def offset(rng: Random): Long
 }
 
 object BinomialScanner {
   private[this] object ProbabilityOneScanner extends BinomialScanner {
-    override def offset(rng: Random): Int = 1
+    override def offset(rng: Random): Long = 1
   }
 
   private[this] class NormalScanner(probability: Double) extends BinomialScanner {
     private[this] val log1p = math.log1p(-probability)
-    override def offset(rng: Random): Int = 1 + (math.log(rng.nextDouble()) / log1p).toInt
+    override def offset(rng: Random): Long = 1 + (math.log(rng.nextDouble()) / log1p).toLong
   }
 
   def apply(probability: Double): BinomialScanner = {
